refactor(plugins-flow-client): move insertBetween logic to a reusable function
